Split message reaction tray from popover
Keep the message action popover below the mobile file-size ratchet by moving
the reaction tray wrapper into its own part file without changing behavior.
